+++ lib/groupdocs/document/option.rb
@@ -1,0 +1,9 @@
+# added in release 2.0.0
+module GroupDocs
+  class Document::TemplateEditorFieldOption < Api::Entity
+    # @attr [String] name
+    attr_accessor :name
+    # @attr [GroupDocs::Document::Rectangle] rect
+    attr_accessor :rect
+  end # Document::TemplateEditorFieldOption
+end # GroupDocs